Enemies now telegraph attacks. Player now has one point of health.

+++ ProtoRunnerV2/src/main/java/com/raggamuffin/protorunnerv2/utils/Vector2.java
@@ -150,7 +150,7 @@ public class Vector2
 	
 	public double GetLength()
 	{
-		return (double)Math.sqrt((I * I) + (J * J));
+		return Math.sqrt((I * I) + (J * J));
 	}
 	
 	public void Output()
